+++ src/main/java/be/ugent/idlab/knows/amo/blocks/nodes/BlankNode.java
@@ -24,9 +24,4 @@ public class BlankNode extends RDFNode {
     public String toString() {
         return "_:" + this.value.toString();
     }
-
-    @Override
-    public String getStringRepr() {
-        return this.toString(); 
-    }
 }